voice framework
